requirements.txt
asgiref==3.7.1
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-speech==2.19.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
mysqlclient==2.1.1
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
python-decouple==3.8
requests==2.31.0
rsa==4.9
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
typing_extensions==4.6.1
urllib3==1.26.16
